Rewrite all preconditions as filtering
The preconditions on multiple arguments were previously rejected by
using `assume_preconditions`. This is not optimal as Hypothesis might
use optimizations to simplify or refine search strategies by analyzing
the source code of the filters and maps.

This patch makes icontract-hypothesis rewrite all preconditions as
`filter` and `map` chains so that Hypothesis can optimize them.