Introduce `inspect` in `test` command
While developing the Vim plug-in, it became apparent that the user is
lost without any feedback. Additionally, ghostwriting is not suitable
for inspecting the tests as ghostwriting should produce valid Python
code, while inspection is expected to give more details about the tests
to the *user*.

This patch introduces `--inspect` to `test` command and enhances the
output in `test`.