Handle `__new__` and forward declarations (#34)
This patch introduces a couple of fixes that handle forward
declarations.

The patch also introduces a couple of fixes to how `__new__` is
handled in this context as we need to pass over a local namespace
including the class corresponding to `__new__` as it is still
not available in the module. Previously, the forward declarations in
`__new__`, which occur quite often!, could not be resolved by
`typing.get_type_hints`.