Pass `self` arugment to contracts
Previously, if a contract involved `self` in its argument, the resulting
Hypothesis strategy for the bound method would have ignored it,
eventually causing a cryptic exception.

This patch adds proper support for `self` in pre-conditions of a bound
method so that `self` is explicitly stated in the initial
`fixed_dictionary` strategy at the begining of the strategy chain.